VIDENG-32 Added collisions systems alongside with damage system, now is a playable game to destroy squares that are coming for you

+++ include/ecs/components/CLife.hpp
@@ -2,7 +2,7 @@
 
 class CLife : public Component {
 public:
-  unsigned int hp = 100;
+  float hp = 100;
   CLife() {};
-  CLife(unsigned int _hp) : hp(_hp) {}
+  CLife(float _hp) : hp(_hp) {}
 };
